// generated/src/app/api-models/site-code-master.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SiteCodeMaster {
  siteCode: string;
  lineOfBusiness: string;
  panelId: string;
  ipaId: string;
  insertProcess: string;
  insertUser: string;
  insertDatetime: Date;
  updateUser: string;
  updateProcess: string;
  updateDatetime: Date;
  description: string;
  sitemUserDefined1: string;
  sitemUserDefined2: string;
  sitemUserDate1: Date;
  sitemUserDate2: Date;
  securityCode: string;
}